requirements.txt
Flask>=2.3.0
Flask-Compress>=1.14
feedparser>=6.0.11
google-generativeai>=0.3.0
python-dotenv>=1.0.0

